requirements.txt
aniso8601
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
coverage==4.5.1
cycler==0.10.0
cypari==2.3.0
Cython==0.28.5
dask==0.19.2
decorator==4.3.0
easydict==1.8
Flask==1.0.2
Flask-RESTful==0.3.6
future==0.16.0
FXrays==1.3.3
idna==2.7
imageio==2.4.1
imgaug==0.2.6
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.12.1
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.2.4
MarkupSafe
matplotlib==2.2.3
networkx==2.1
nose==1.3.7
numpy==1.15.1
opencv-python
opencv-contrib-python
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
plink==2.2
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
pypng==0.0.18
PyQt5==5.11.2
PyQt5-sip==4.19.12
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
spherogram==1.8.1
toolz==0.9.0
http://download.pytorch.org/whl/cu80/torch-0.4.1-cp35-cp35m-linux_x86_64.whl
torchvision
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.23
visdom==0.1.8.5
wcwidth==0.1.7
websocket-client==0.51.0
Werkzeug==0.14.1
